Add nav toggle js
The sidebar is naturally collapsed and only the icons are visible (they all have aria-label); HOWEVER js checks with media queries the screen size on load and resize so that if there is large screen the sidebar expands, and if there's small screen it collapses.
It is also possible of course to toggle the navigation open and close using the button, it shows an open/close book when it's open/close.

ISSUES: the transition between open and close is a little abrupt, will try to make a smoother transition.