Create initial book template from codrops
Created a book that flips to its back where there is  summary and enables view inside where there is a quote from the book. The book appearance and general code is taken from codrops:
https://tympanus.net/codrops/2013/01/08/3d-book-showcase/

This version is created using js with classes for the books.
The appearance and size are not customized to my choice yet, this is  only the initial code.

IMPORTANT:
add credit to codrops for inspiring and heavily helping with this code.