Add disabling of arrows at post-it extremities
When scrolling through the announcements the thumbnails with arrows at the extremieties it's best practice to disable the arrows. This was achieved here using a mutation observer to check the class change on the first and last thumbnail, if the class is highlighted that means that there is no more to scroll in that direction so the arrow is disabled.
I used this website for general idea of how to build the mutation observer:
https://www.seanmcp.com/articles/event-listener-for-class-change/

and in order to add parameters to the mutation observer callback function (this way building one function instead of two) I used this website:
https://medium.com/@estebanfrs/mutationobserver-api-additional-params-186b86389148

TBC:
- style the arrows