Style and move post it arrows for better use
The arrows should be at a location where they are easy to use, not yet certain if this position with this design is a good idea since it covers the post it a little.
the design is like a post it arrow, with colors to go with it.
HTML: since the arrows are not at the bottom but the thumbnails are, they didn't need to be in a div together, this changed slightly the html.
JS: changes to the html lead to slight changes in js, nothing important.

NOTE:
the js on which the scrolling and thumbnails are based is from the following source. made slight changes and added buttons for scrolling.
https://www.30secondsofcode.org/css/s/horizontal-gallery