Add tags and links to books
The links are in the more-info buttons, and lead nowhere presently.
The tags are not strictly necessary but they now decide the colors of the books, to make it more 'uniform'. There is also a default color.
If this ever becomes more functional the book info will come from sql and the tags as classes will not be necessary.