Fix overflow x-axis on small screen
At small screen, when opening the navbar it caused the x-axis to overflow (because of min-width of the aside-main). To prevent this added a width of 100% to the flex container, and a max-width of 100vw. This makes sure that there is place for the scrollbar but not beyond it. also an overflow clip so you cannot access what's pushed to the side.
Also because this is a flex container, the navbar didn't expand to its full width (to make more space for the container), to fix this added a min-width of same size as width.