Fix typeError on click to preview recommendation
The problem was not with the book recommendation modal, it was with the preview. Since the preview was only the front of the book and nothing else it caused a typeError in the event listeners (they were trying to work on click to perform their functions but could not because there were components 'missing'). I excluded the book preview by adding :not(.preview) within the query selector, this solved the problem.
NOTE:
- even if the browser doesn't accept :not() this should still work, as stated in previous commit, the typeError is displayed in console but everything still works fine. As it were this should be supported in most browsers.