Make navbar transition smooth
Changed the navbar default setting to be for computer, not mobile (since it's the default for the rest). Also changed the nav list items from flex to grid for better control over the column sizes and so that the changes can be smaller and transitioned better. Finally wanted to make the page names transition smoothly from visibly to invisible. used transition opacity for this on the default page and transition opacity of 0s on the collapsed page, this makes it so that when collapsing there is no transition but when resuming expanded navbar there is a transition, all together the effect is smoother.