Fix sidebar transition to smooth
The sidebar would transition in width but the inside changed from display none to regular, making the transition jumpy. to fix this instead of display none, when collapsed in big screen the inside is moved on the x axis until it's not longer visible. note that because the left arrow on the announcements part is made to move to the other side it was addressed individually so that it would move smoothly with the rest.