Change styling, dark mode
Chose the general styling for the site - woody christmasy, a cozy feeling of wood and fireplaces that is often associated with books. This theme will remain for both dark and light themes but the color scheme will naturally change a little.

All the color contrasts (except for the books which need to be changed) have been tested for accessability using:
https://webaim.org/resources/contrastchecker/

there is still a lot to do, fixing the colors, book colors, fonts, icon and arrow colors, shadows, etc for the dark mode. and then fixing it for the light mode, and creating the theme toggle functionality.