Change colors for dark theme + generalize
The dark theme is nearly finishied, left only to choose fonts (which will be for both themes of course) and possibly final tweaking of clrs later on. Generalized some clr changes and filter inversions for easier manipulation of themes.
NOTE: the icons are changed from black to white using invert(), they could be changed to a precise color using filters or by implanting the full svg in the html and manipulating the svg. for the sake of less code i chose to implant images instead of the svg code, remains to be seen which is the better choice.